+++ coreblocks/frontend/decoder/instr_description.py
